+++ libs/utils/src/array-to-object.ts
@@ -1,5 +1,9 @@
-export function ArrayToObject<T>(array: T[], columns: string[]) {
-  const obj = {};
+interface ParsedObjectInterface {
+  [key: string]: unknown;
+}
+
+export function ArrayToObject<T>(array: T[], columns: string[]): ParsedObjectInterface {
+  const obj: ParsedObjectInterface = {};
   for (let i = 0; i < columns.length; i++) {
     obj[columns[i]] = array[i];
   }
