+++ libs/utils/src/index.ts
@@ -1,2 +1,3 @@
 export * from './array-to-object';
 export * from './products-index-name';
+export * from './batch-action';
